src/py: organized experiments under new "experiments" module. Added filtering support to SampleSet.
All experiments need to be organized in the "experiments" module with a single integer id. This makes running and re-running greatly simplified. The SampleSet class used to create random samples of training/testing data was extended to add basic filtering support. Many of the experiments want to only use a few classes or environments. This logic enables that.

Additionally, the main scripts were improved to make running experiments more efficient. The main run.sh is now only used for running experiments. This addresses Objectives 1 & 2. The live testing implementation is now started by a different script: listen.sh. This is the script targeting Objective 3.