src/py: added multi template convolution feature.
This required adding a class_rep method to SampleSet so that a representation of each class type could be easily retrieved. Unit tests were added for this. The feature classes were revised to have a "generate" method that takes a TestSample - this made it possible to construct a Feature object that can be used to generate features for each TestSample.

Demo: Try the new template feature with the test runner: ./run.sh M Y m